Move code.py functionality into lib
